+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changes in 0.2
 ==============
 
  * Basic support for numbers 100-999
+ * Stop overlap of 10 blocks
 
 Changes in [0.1](https://github.com/amalon/minetest-numeracy/releases/tag/v0.1) (2021-04-04)
 ============================================================================================